requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyyaml
requests
torch
seaborn
tabulate
mlflow
skops
